+++ src/gui/main_window.py
@@ -54,6 +54,7 @@ from src.platforms.fetlife import FetLifePlatform
 from src.platforms.instagram import InstagramPlatform
 from src.platforms.onlyfans import OnlyFansPlatform
 from src.platforms.snapchat import SnapchatPlatform
+from src.platforms.threads import ThreadsPlatform
 from src.platforms.twitter import TwitterPlatform
 from src.utils.constants import (
     APP_NAME,
@@ -359,6 +360,7 @@ class MainWindow(QMainWindow):
         'onlyfans': lambda am, aid, pn: OnlyFansPlatform(account_id=aid, profile_name=pn),
         'fansly': lambda am, aid, pn: FanslyPlatform(account_id=aid, profile_name=pn),
         'fetlife': lambda am, aid, pn: FetLifePlatform(account_id=aid, profile_name=pn),
+        'threads': lambda am, aid, pn: ThreadsPlatform(account_id=aid, profile_name=pn),
     }
 
     def _build_platforms(self):